Refactor codebase: migrate from JavaScript to TypeScript, remove unused files, and update imports
