+++ index.html
@@ -19,7 +19,7 @@
 	</div>
 
 	<script defer type="module">
-		import NianceEditor from "./lib/main.js";
+		import NianceEditor from "./dist/niance-editor.es.js";
 
 		const {CreateEditor, InitializeEditor} = NianceEditor;
