--- lib/components/SuggestionNavigation.js
@@ -1,5 +1,0 @@
-export const SuggestionNavigationProps = {
-	firstSuggestionIndex: 0,
-	currentSuggestionIndex: 0,
-	lastSuggestionIndex: 0,
-};
