+++ lib/components/SuggestionNavigation.ts
@@ -1,0 +1,12 @@
+interface SuggestionNavigationProps {
+    firstSuggestionIndex: number;
+    currentSuggestionIndex: number;
+    lastSuggestionIndex: number;
+  }
+  
+  export const SuggestionNavigationProps: SuggestionNavigationProps = {
+    firstSuggestionIndex: 0,
+    currentSuggestionIndex: 0,
+    lastSuggestionIndex: 0,
+  };
+
