--- lib/global/DOMElements.js
@@ -1,6 +1,0 @@
-const DOMElements = {
-	editor: undefined,
-	suggestionContainer: undefined,
-}
-
-export default DOMElements;
