+++ lib/global/DOMElements.ts
@@ -1,0 +1,7 @@
+const DOMElements: { editor?: HTMLElement; suggestionContainer?: HTMLElement } = {
+    editor: undefined,
+    suggestionContainer: undefined,
+  };
+  
+  export default DOMElements;
+
