--- lib/main.js
@@ -1,8 +1,0 @@
-import {CreateEditor, InitializeEditor} from "./editor/editor.js";
-
-const NianceEditor = {
-	CreateEditor,
-	InitializeEditor,
-}
-
-export default NianceEditor;
