+++ tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./lib/main.ts","./lib/components/caret.ts","./lib/components/completion.ts","./lib/components/suggestionnavigation.ts","./lib/components/syntaxhighlighting.ts","./lib/core/fuzzysearch.ts","./lib/core/lexer.ts","./lib/core/utility.ts","./lib/editor/editor.ts","./lib/global/domelements.ts","./lib/global/global.ts","./lib/secrets/veryimportantfiledonottouchpleasecontainssecrettouniverse.ts"],"version":"5.7.3"}
