+++ vite.config.cjs
@@ -4,7 +4,7 @@ const {defineConfig} = require('vite')
 module.exports = defineConfig({
 	build: {
 		lib: {
-			entry: path.resolve(__dirname, 'lib/main.js'),
+			entry: path.resolve(__dirname, 'built/main.js'),
 			name: 'niance-editor',
 			fileName: (format) => `niance-editor.${format}.js`
 		}
